+++ R/obr-package.R
@@ -7,4 +7,6 @@
 #' @concept national debt
 #' @concept UK budget
 #' @concept economic forecasts
+#' @importFrom utils head packageVersion str
+#' @importFrom stats aggregate
 "_PACKAGE"